+++ Package.swift
@@ -1,11 +1,11 @@
-// swift-tools-version:5.6
+// swift-tools-version:5.8
 import PackageDescription
 
 let package = Package(
     name: "Assist",
 //    platforms: [.macOS(.v10_15), .macCatalyst(.v13), .iOS(.v13), .tvOS(.v13), .watchOS(.v6)],
     platforms: [
-      .macOS(.v12)
+      .macOS(.v13)
 //        .iOS(.v11),
 //        .watchOS(.v5),
 //        .tvOS(.v11)
@@ -17,8 +17,8 @@ let package = Package(
     ],
     dependencies: [
         .package(url: "https://github.com/blackjacx/Engine", from: "0.0.3"),
-        .package(url: "https://github.com/blackjacx/ASCKit", from: "0.1.0"),
-        .package(url: "https://github.com/apple/swift-argument-parser", from: "1.1.1"),
+        .package(url: "https://github.com/blackjacx/ASCKit", from: "0.2.0"),
+        .package(url: "https://github.com/apple/swift-argument-parser", from: "1.2.2"),
         .package(url: "https://github.com/vapor/jwt-kit.git", from: "4.1.0"),
         .package(url: "https://github.com/kareman/SwiftShell", from: "5.1.0")
     ],