+++ Sources/Core/Shell/Xcodebuild.swift
@@ -9,37 +9,54 @@ import Foundation
 import SwiftShell
 
 public struct Xcodebuild {
-}
 
-public extension Xcodebuild {
+    public static func execute(subcommand: Subcommand,
+                               deviceIds: [String],
+                               derivedDataUrl: URL) throws {
+        let destinations = deviceIds
+            .map { "platform=iOS Simulator,id=\($0)" }
+            .flatMap { ["-destination", $0] }
+        let commonArgs = ["xcodebuild", subcommand.name, "-derivedDataPath", derivedDataUrl.path()] + destinations
+        let invocations: [Invocation]
 
-    enum Command: String {
-        case buildForTesting = "build-for-testing"
-        case testWithoutBuilding = "test-without-building"
+        switch subcommand {
+        case let .buildForTesting(workspace, schemes):
+            invocations = schemes.map {
+                Invocation(arguments: commonArgs + ["-workspace", workspace, "-scheme", $0])
+            }
+
+        case let .testWithoutBuilding(xcTestRunFile, resultsBundleURL):
+            var args = ["-xctestrun", xcTestRunFile.path()]
+            if let resultsBundleURL {
+                args += ["-resultBundlePath", resultsBundleURL.path]
+            }
+            invocations = [Invocation(arguments: commonArgs + args)]
+        }
+
+        for invocation in invocations {
+            try invocation.execute()
+        }
     }
-    
-    static func execute(cmd: Command,
-                        workspace: String,
-                        schemes: [String],
-                        deviceIds: [String],
-                        testPlan: String? = nil,
-                        resultsBundleURL: URL? = nil) throws {
 
-        let destinations = deviceIds.map { "platform=iOS Simulator,id=\($0)" }
+    // MARK: - Subcommand
 
-        for scheme in schemes {
-            var args = ["xcodebuild", cmd.rawValue, "-workspace", workspace, "-scheme", scheme]
-            destinations.forEach { args += ["-destination", $0] }
+    public enum Subcommand {
+        case buildForTesting(workspace: String, schemes: [String])
+        case testWithoutBuilding(xcTestRunFile: URL, resultsBundleURL: URL? = nil)
 
-            if let testPlan = testPlan {
-                args += ["-testPlan", testPlan]
+        var name: String {
+            switch self {
+            case .buildForTesting: return "build-for-testing"
+            case .testWithoutBuilding: return "test-without-building"
             }
+        }
+    }
 
-            if let resultBundleURL = resultsBundleURL {
-                args += ["-resultBundlePath", resultBundleURL.path]
-            }
+    private struct Invocation {
+        let arguments: [String]
 
-            let out = run("xcrun", args)
+        func execute() throws {
+            let out = run("xcrun", arguments)
 
             if let error = out.error {
                 Logger.shared.error(out.stderror)