+++ Sources/Snap/commands/Snap.swift
@@ -58,13 +58,13 @@ public final class Snap: ParsableCommand {
     var schemes: [String]
 
     @Option(help: "The name of the TestPlan running the screenshot tests.")
-    var testPlanName: String?
+    var testPlanName: String
 
     @Option(parsing: .upToNextOption, help: "The appearances the screenshots should be made for, e.g. --appearances \(Simctl.Style.allCases.map({"\"\($0.parameterName)\""}).joined(separator: " "))")
     var appearances: [Simctl.Style] = [.light]
 
-    @Option(parsing: .upToNextOption, help: "Devices you want to generate screenshots for, e.g. --devices \(Simctl.DeviceType.allCases.map({"\"\($0.parameterName)\""}).joined(separator: " "))")
-    var devices: [Simctl.DeviceType] = [.iPhone12Pro]
+    @Option(parsing: .upToNextOption, help: "Devices you want to generate screenshots for (run `xcrun simctl list` to list all possible devices)")
+    var devices: [String] = ["iPhone 14 Pro"]
 
     @Option(help: "The destination directory where the screenshots and the zip archive should be stored.")
     var destinationDir: String?
@@ -72,8 +72,8 @@ public final class Snap: ParsableCommand {
     @Option(help: "The zip file name that should be used.")
     var zipFileName: String
 
-    @Option(help: "An optional platform to be used. Omit to use the latest. Currently only iOS is supported.")
-    var platform: Simctl.Platform?
+    @Option(help: "An optional runtime to be used. Omit to use the latest.")
+    var runtime: String?
 
     public init() {
         
@@ -88,65 +88,70 @@ public final class Snap: ParsableCommand {
             throw ValidationError("No target specified.")
         }
 
-        if let destinationDir = self.destinationDir {
+        if let destinationDir {
             var isDir: ObjCBool = false
             guard FileManager.default.fileExists(atPath: destinationDir, isDirectory: &isDir), isDir.boolValue else {
                 throw ValidationError("\(destinationDir) does not exist or is no directory.")
             }
         }
 
-        if let platform = self.platform {
-            guard try Simctl.isPlatformValid(platform.rawValue) else {
-                throw ValidationError("\(platform) not installed on your system. Run `xcrun simctl list` to find available ones.")
+        if let runtime {
+            guard try Simctl.isRuntimeNameValid(runtime) else {
+                let availableRuntimes = ListFormatter.localizedString(byJoining: try Simctl.availableRuntimes().map(\.name))
+                throw ValidationError("\(runtime) not installed on your system. Valid runtimes: \(availableRuntimes).")
             }
         }
     }
 
     public func run() throws {
         let outURL: URL
+        let derivedDataUrl = try FileManager.createTemporaryDirectory()
 
-        if let destinationDir = destinationDir {
+        if let destinationDir {
             outURL = URL(fileURLWithPath: destinationDir, isDirectory: true)
         } else {
             outURL = try FileManager.createTemporaryDirectory()
         }
 
         do {
-            let platform = try self.platform?.rawValue ?? Simctl.latestAvailableIOS()
+            let runtimeName = try self.runtime ?? Simctl.latestAvailableIosRuntime()
 
             let configMessage = """
                 Using the following config:
-                    styles: \(appearances.map { $0.parameterName })
-                    devices: \(devices.map { $0.parameterName })
-                    platform: \(platform)
-                    schemes: \(schemes)
+                    styles: \(ListFormatter.localizedString(byJoining: appearances.map { $0.parameterName }))
+                    devices: \(ListFormatter.localizedString(byJoining: devices))
+                    platform: \(runtimeName)
+                    schemes: \(ListFormatter.localizedString(byJoining: schemes))
                     destination: \(outURL.path.appendPathComponent(zipFileName))
                 """
             Logger.shared.info(configMessage)
 
             Simctl.killAllSimulators()
 
-            Logger.shared.info("Finding runtime for platform \(platform)")
-            let runtime = try Simctl.runtimeForPlatform(platform)
+            Logger.shared.info("Finding runtime for platform \(runtimeName)")
+            let runtime = try Simctl.runtime(for: runtimeName)
             Logger.shared.info("Runtime found \(runtime)")
 
             Logger.shared.info("Find IDs of preferred device IDs")
-            let deviceIds = try Simctl.deviceIdsFor(deviceTypes: devices, runtime: runtime)
+            let deviceIds = try Simctl.deviceIdsFor(deviceNames: devices, runtime: runtime)
             Logger.shared.info("Device IDs Found: \(deviceIds)")
 
             Logger.shared.info("Building all requested schemes for testing")
-            try Xcodebuild.execute(cmd: .buildForTesting,
-                                   workspace: workspace,
-                                   schemes: schemes,
-                                   deviceIds: deviceIds)
+            try Xcodebuild.execute(subcommand: .buildForTesting(workspace: workspace, schemes: schemes),
+                                   deviceIds: deviceIds,
+                                   derivedDataUrl: derivedDataUrl)
 
-            Logger.shared.info("Taking screenshots for all requested configs")
+            Logger.shared.info("Taking screenshots for all requested configurations")
             try Simctl.snap(styles: appearances,
                             workspace: workspace,
                             schemes: schemes,
+                            derivedDataUrl: derivedDataUrl,
                             testPlanName: testPlanName,
+                            runtime: runtimeName.components(separatedBy: " ").last!, // grab the version number of the provided runtime
+                            arch: ProcessInfo.machineHardwareName,
+                            platform: "iphonesimulator", // we're always generating on a simulator
                             deviceIds: deviceIds,
-                            outURL: outURL,
+                            outUrl: outURL,
                             zipFileName: zipFileName)
 
             Logger.shared.info("Find your screens in \(outURL.path)")
@@ -175,5 +180,3 @@ struct Options: ParsableArguments {
 }
 
 extension Simctl.Style: ExpressibleByArgument {}
-extension Simctl.DeviceType: ExpressibleByArgument {}
-extension Simctl.Platform: ExpressibleByArgument {}